Fix `pf login` error when called without params
When calling `pf login` without any parameters `$username` variable was NULL
also if there weren't any authentificated user `$phpfog->username()`
was returning also NULL (with notice) so the equality
`$phpfog->username() == $username` was true and displayed message was
`Already logged in as .` but no one was authentificated.

Now if `$username` is emtpy it promts for an username and then does it's job
accordingly.

Also changed `$phpfog->username()` method so it doesn't throw any notice.